Move byte buffer String conversion logic to JavaUtils
